# packages/detensor-cli/detensor_cli-0.1.9.tar.gz/detensor_cli-0.1.9/detensor_cli/template/pro_project/storage-tt/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "storage API是储存层的对象储存API",
        "title": "storage API",
        "contact": {
            "name": "bravomikekilo",
            "email": "bravomikekilo@buaa.edu.cn"
        },
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/core/v1/function": {
            "get": {
                "description": "按照请求创建私有函数, 返回创建的私有函数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Function 私有函数"
                ],
                "summary": "查找私有函数对象 API",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "限制私有函数对象要属于某个数据声明",
                        "name": "declare",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制私有函数对象要被某个计算合约使用",
                        "name": "contract",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "查找出的私有函数对象",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/objects.Function"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "按照请求创建私有函数, 返回创建的私有函数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Function 私有函数"
                ],
                "summary": "创建私有函数API",
                "parameters": [
                    {
                        "description": "待创建的私有函数对象",
                        "name": "function",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/objects.Function"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "私有函数对象创建成功",
                        "schema": {
                            "$ref": "#/definitions/objects.Function"
                        }
                    }
                }
            }
        },
        "/core/v1/function/{name}": {
            "get": {
                "description": "根据私有函数对象的名称,在本地数据库中查找私有函数对象",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Function 私有函数"
                ],
                "summary": "按照名称查找私有函数对象 API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "私有函数对象的名称或uid",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "查找出的私有函数对象",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/objects.Function"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "按照请求更改私有函数, 返回更改后的私有函数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Function 私有函数"
                ],
                "summary": "更改私有函数API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要更新的私有函数对象的名称或uid",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "要更新的新私有函数对象",
                        "name": "newFunc",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/objects.Function"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功, 返回更新后的对象",
                        "schema": {
                            "$ref": "#/definitions/objects.Function"
                        }
                    },
                    "401": {
                        "description": "对象格式有误",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "找不到对应的私有对象"
                    }
                }
            },
            "delete": {
                "description": "按照请求删除私有函数, 返回删除的私有函数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Function 私有函数"
                ],
                "summary": "删除私有函数 API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要删除的私有函数对象的名称, 或是UID",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功, 返回删除的对象",
                        "schema": {
                            "$ref": "#/definitions/objects.Function"
                        }
                    },
                    "404": {
                        "description": "找不到对应的私有对象"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "objects.Function": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/objects.Metadata"
                },
                "spec": {
                    "$ref": "#/definitions/objects.FunctionSpec"
                },
                "status": {
                    "$ref": "#/definitions/objects.FunctionStatus"
                }
            }
        },
...
        "objects.FunctionSpec": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "objects.FunctionStatus": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                }
            }
        },
        "objects.Metadata": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "uid": {
                    "type": "integer"
                }
            }
        }
    }
}